requirements.txt
nicegui==1.5.6
gspread==6.1.2
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httplib2==0.22.0
oauth2client==4.1.3
requests==2.32.3
email-validator==2.1.0.post1
